Refactor repository structure and update docs
