+++ docs/game_logic_module.md
@@ -1,0 +1,15 @@
+# Game Logic Module
+
+Conceptual design for RPG-specific mechanics and systems.
+
+## Planned Components
+- `quest_system.py` – quest management and progression
+- `dialogue_system.py` – character interaction and conversation trees
+- `tagging.py` – decision processing and trait tagging
+- `narrative_manager.py` – story branching and flow control
+- `character_manager.py` – player and NPC management
+
+## Integration Points
+- Interfaces with psychology module for decision tracking
+- Uses core engine for rendering and state management
+- Loads content from data directory
