+++ docs/psychology_profiling_module.md
@@ -1,0 +1,23 @@
+# Psychology Profiling Module
+
+Conceptual design for covert psychological assessment that analyzes player decisions.
+
+## Planned Components
+- `traits.py` – personality trait definitions
+- `scoring_engine.py` – decision-to-trait mapping algorithms
+- `profile_generator.py` – comprehensive personality profile creation
+- `behavioral_analyzer.py` – pattern recognition in player choices
+- `chronicle_writer.py` – Hero's Chronicle narrative generation
+
+## Ethical Considerations
+- Transparent data usage policies
+- Opt-in psychological profiling
+- Clear explanation of assessment methods
+- Privacy-first design
+
+## Supported Assessments
+- Big Five personality traits
+- Risk tolerance assessment
+- Attachment style evaluation
+- Decision-making patterns
+- Moral reasoning analysis
