+++ docs/utility_functions.md
@@ -1,0 +1,10 @@
+# Utility Functions
+
+Conceptual list of shared helpers planned for the Janus project.
+
+## Planned Components
+- `config.py` – configuration management
+- `logging.py` – logging and debugging utilities
+- `data_loader.py` – JSON/YAML data loading helpers
+- `validation.py` – input validation and sanitization
+- `text_processing.py` – text formatting and manipulation
