+++ docs/vision_hamartia_system.md
