--- src/game/README.md
@@ -1,17 +1,0 @@
-# Game Logic Module
-
-This module handles the RPG-specific game mechanics and systems.
-
-## Components
-
-- **quest_system.py**: Quest management and progression
-- **dialogue_system.py**: Character interaction and conversation trees
-- **choice_engine.py**: Decision processing and consequence handling
-- **narrative_manager.py**: Story branching and flow control
-- **character_manager.py**: Player and NPC management
-
-## Integration Points
-
-- Interfaces with psychology module for decision tracking
-- Uses core engine for rendering and state management
-- Loads content from data directory
