+++ src/modules/README.md
@@ -1,0 +1,8 @@
+# Engine Modules
+
+Reusable components supporting the core engine.
+
+## Available Modules
+- `tagging.py` – Trait tagging utilities for player choices.
+- `save_system.py` – JSON-based save/load helpers.
+- `telemetry.py` – Lightweight gameplay event logging.
