--- src/psychology/README.md
@@ -1,26 +1,0 @@
-# Psychology Profiling Module
-
-This module implements the covert psychological assessment system that analyzes player decisions.
-
-## Components
-
-- **traits.py**: Personality trait definitions (Big Five, etc.)
-- **scoring_engine.py**: Decision-to-trait mapping algorithms
-- **profile_generator.py**: Comprehensive personality profile creation
-- **behavioral_analyzer.py**: Pattern recognition in player choices
-- **chronicle_writer.py**: Hero's Chronicle narrative generation
-
-## Ethical Considerations
-
-- Transparent data usage policies
-- Opt-in psychological profiling
-- Clear explanation of assessment methods
-- Privacy-first design
-
-## Supported Assessments
-
-- Big Five Personality Traits
-- Risk tolerance assessment
-- Attachment style evaluation
-- Decision-making patterns
-- Moral reasoning analysis
