--- src/utils/README.md
@@ -1,11 +1,0 @@
-# Utility Functions
-
-Common utilities and helper functions used across the Janus project.
-
-## Components
-
-- **config.py**: Configuration management
-- **logging.py**: Logging and debugging utilities
-- **data_loader.py**: JSON/YAML data loading helpers
-- **validation.py**: Input validation and sanitization
-- **text_processing.py**: Text formatting and manipulation
