+++ tests/README.md
@@ -1,33 +1,8 @@
 # Testing Suite
 
-Comprehensive testing for all aspects of the Janus project.
-
-## Test Categories
-
-### Unit Tests (`unit/`)
-- Individual component testing
-- Psychology algorithm validation
-- Core engine functionality
-- Utility function verification
-
-### Integration Tests (`integration/`)
-- End-to-end gameplay scenarios
-- Psychology profiling accuracy
-- Cross-module interactions
-- Data loading and validation
-
-## Testing Guidelines
-- Test psychological assessment accuracy
-- Validate ethical considerations
-- Ensure gameplay balance
-- Performance benchmarking
+Placeholder for future tests of the Janus project.
 
 ## Running Tests
 ```bash
-# Run all tests
-python -m pytest tests/
-
-# Run specific test category
-python -m pytest tests/unit/
-python -m pytest tests/integration/
+python -m pytest
 ```
